+++ src/main/java/com/ideality/coreflow/attachment/command/application/controller/AttachmentConrtoller.java
@@ -1,0 +1,4 @@
+package com.ideality.coreflow.attachment.command.application.controller;
+
+public class AttachmentConrtoller {
+}
